pyproject.toml
[tool.poetry]
name = "jupyter-snetwork"
version = "0.0.5"
description = ""
authors = [
    "Marcelo Hashimoto <marcelo.hashimoto@gmail.com>"
]
readme = "README.md"
repository = "https://github.com/hashiprobr/jupyter-snetwork"
packages = [
    { include = "snetwork" }
]

[tool.poetry.dependencies]
python = "^3.11"
jupyter-jchannel = "~0.3.7"

[tool.poetry.group.dev.dependencies]
nbclassic = "^1.2.0"
notebook = "^7.3.2"
jupyterlab = "^4.3.4"
ipywidgets = "^8.1.5"
autopep8 = "^2.3.2"
pytest = "^8.3.4"
pytest-asyncio = "~0.25.2"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
sphinx = "^8.1.3"

[tool.poetry.scripts]
nbclassic = 'scripts:nbclassic'
notebook = 'scripts:notebook'
lab = 'scripts:lab'
tests = 'scripts:tests'
docs = 'scripts:docs'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
